Do some trivial changes to fix warnings
