+++ sdk/signalr/Microsoft.Azure.SignalR.Serverless.Protocols/src/Messages/OpenConnectionMessage.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.Azure.SignalR.Serverless.Protocols
+{
+    /// <summary>
+    /// A message represents that a connection is opened.
+    /// </summary>
+    public class OpenConnectionMessage : ServerlessMessage
+    {
+    }
+}
