+++ sdk/signalr/Microsoft.Azure.SignalR.Serverless.Protocols/src/Microsoft.Azure.SignalR.Serverless.Protocols.csproj
@@ -5,7 +5,7 @@
     <PackageId>Microsoft.Azure.SignalR.Serverless.Protocols</PackageId>
     <Version>1.7.0-beta.1</Version>
     <SignAssembly>true</SignAssembly>
-    <NoWarn>AZC0001;CA2000;CA2208;AZC0100;CA2215;SA1402;SA1649;</NoWarn>
+    <NoWarn>AZC0001</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
